Day 3 - Turtle Program Sample Solution
Here is a sample solution for the final task in day 3. Note that this is just a sample solution, and you're free to do whatever you want with your program, as long as it fits the task requirements. Pay attention to the use of conditionals to check user input to be valid. Although this is not mandatory, we recommend you do this as good practice, as annoying as this can be.